Keep provider maintenance off the app boot path
