+++ packages/provider-bridge-protocol/src/index.ts
@@ -14,6 +14,7 @@
 export * from "./version.js";
 export * from "./handshake.js";
 export * from "./execution-options.js";
+export * from "./provider-maintenance.js";
 export * from "./requests.js";
 export * from "./notifications.js";
 export * from "./bridge-requests.js";
